Narrowing return exit speeds observations

+++ src/Workers/MiningOptimization/ReturnPositionObservations.cpp
@@ -13,7 +13,7 @@ namespace WorkerMiningOptimization
         // - High exit speed saves 4 frames
         // - Medium exit speed saves 2 frames
         // - Collisions cost an extra order process timer cycle
-        uint32_t total = collision + lowExitSpeed + mediumExitSpeed + highExitSpeed;
+        uint16_t total = collision + lowExitSpeed + mediumExitSpeed + highExitSpeed;
         if (total == 0) return 0.0;
 
         return (double)(((int)collision * 9) - ((int)mediumExitSpeed * 2) - ((int)highExitSpeed * 4)) / (double)total;
